// apps/backend/src/common/uploaded-file.type.ts
export type UploadedFile = {
  fieldname: string;
  originalname: string;
  encoding: string;
  mimetype: string;
  size: number;
  destination?: string;
  filename?: string;
  path?: string;
  buffer: Buffer;
};

export type UploadedFileFields = {
  image?: UploadedFile[];
  backImage?: UploadedFile[];
};